refactor: set code hash in KakarotAccount::new
