+++ crates/ef-testing/src/evm_sequencer/evm_state/v1.rs
@@ -388,8 +388,14 @@ mod tests {
         ]);
 
         // When
-        let account =
-            KakarotAccount::new(&TEST_CONTRACT_ADDRESS, &bytecode, U256::ZERO, &[]).unwrap();
+        let account = KakarotAccount::new(
+            &TEST_CONTRACT_ADDRESS,
+            &bytecode,
+            U256::ZERO,
+            U256::ZERO,
+            &[],
+        )
+        .unwrap();
         sequencer.setup_account(account).unwrap();
 
         // Then
@@ -438,10 +444,12 @@ mod tests {
             &TEST_CONTRACT_ADDRESS,
             &contract_bytecode,
             contract_nonce,
+            U256::ZERO,
             &[],
         )
         .unwrap();
-        let eoa = KakarotAccount::new(&PUBLIC_KEY, &Bytes::default(), eoa_nonce, &[]).unwrap();
+        let eoa = KakarotAccount::new(&PUBLIC_KEY, &Bytes::default(), eoa_nonce, U256::ZERO, &[])
+            .unwrap();
         sequencer.setup_account(contract).unwrap();
         sequencer.setup_account(eoa).unwrap();
         let execution_result = sequencer.execute_transaction(transaction);
